link to lesson and move page to proper dir level
