+++ app/solidity_lessons/page.tsx
@@ -1,38 +1,10 @@
-import CourseViewer, { ICourseViewer } from "../common_components/course_viewer/courseViewer";
-
-const data: ICourseViewer = {
-  courseName: "Solidity Basics",
-  courseDescription: "Learn the basics of Solidity blah blah blah Enim similique quaerat id nam voluptatemEnim similique quaerat id nam voluptatemEnim similique quaerat id nam voluptatemEnim similique quaerat id nam voluptatemvoluptatemEnim similique quaerat id nam voluvoluptatemEnim similique quaerat id nam voluvoluptatemEnim similique quaerat id nam volu",
-  courseThumbnail: "https://i.imgur.com/Jvh1OQm.jpeg",
-  prerequisites: ["Basic Programming", "Basic Javascript"],
-  lessons: [
-    { 
-      lessonName: "Basic Datatypes",
-      lessonDescription: "Learn about the basic datatypes in Solidity Sint eveniet id rerum. Enim similique quaerat id nam voluptatem. ",
-      thumbnailPath: "https://i.imgur.com/3ZQc5nJ.png"
-   }, {
-      lessonName: "Maps and Structs",
-      lessonDescription: "Learn about map datastructure in Solidity Sint eveniet id rerum. Enim similique quaerat id nam voluptatem.",
-      thumbnailPath: "https://i.imgur.com/3ZQc5nJ.png"
-   },{
-      lessonName: "Functions",
-      lessonDescription: "Learn about functions in Solidity Sint eveniet id rerum. Enim similique quaerat id nam voluptatem.",
-      thumbnailPath: "https://i.imgur.com/3ZQc5nJ.png"
-   }
-
-  ]
-}
+//component to browse courses
 
 const Page = () => {
   return (
-    <CourseViewer 
-      courseName={data.courseName} 
-      courseDescription={data.courseDescription} 
-      courseThumbnail={data.courseThumbnail}
-      prerequisites={data.prerequisites}
-      lessons={data.lessons} 
-    />
+    <div>
+      test
+    </div>
   )
 }
-
 export default Page;
